Moves demas plot.
